Add files via upload
This is a series of functions that is intended to identify errors in the conversion of a pdf file to text using tesseract OCR.  Some preliminary cleaning of the text files is performed in order to strip out punctuation, stop words, two letter words and some common errors that tesseract made that were more common.  The error checker function will compare each token of a text to the NLTK Wordnet Dictionary to identify errors. Note that the author realizes that this would not catch an error in which the original word was 'cat' and tesseract output 'hat.  

The output can be selected by the user.  You can retrieve the  percentage error per document, which can helpful when dealing with hundreds of documents that have been converted.  In addition, you can choose to output the frequency distribution of errors across n text files (as many as you want to pass to the function) or receive the output in the form of a bar graph. 

As of the date of this posting the code is not yet finalized. I still need to clean it up and create the final pipeline to traverse a directory and pipe the files through the functions.